pipeline: move import of pipeline out of the root of the library before every import of entropylab.flame would import from entropylab __init__.py whole pipeline, slowing down significantly flame because of thick library stack of pipeline. Now this is moved to pipeline __init__.py speeding up flame significantly.
